type error corrected4

+++ components/form/delete-site-form.tsx
@@ -11,23 +11,26 @@ import va from "@vercel/analytics";
 export default function DeleteSiteForm({ siteName }: { siteName: string }) {
   const { id } = useParams() as { id: string };
   const router = useRouter();
+
+  const deleteSiteData = (data: FormData) => {
+    window.confirm("Are you sure you want to delete your site?") &&
+      deleteSite(data, id, "delete")
+        .then(async (res) => {
+          if (res.error) {
+            toast.error(res.error);
+          } else {
+            va.track("Deleted Site");
+            router.refresh();
+            router.push("/sites");
+            toast.success(`Successfully deleted site!`);
+          }
+        }).catch((err: Error) => toast.error(err.message))
+  }
+
+
   return (
     <form
-      action={async (data: FormData) =>
-        window.confirm("Are you sure you want to delete your site?") &&
-        deleteSite(data, id, "delete")
-          .then(async (res) => {
-            if (res.error) {
-              toast.error(res.error);
-            } else {
-              va.track("Deleted Site");
-              router.refresh();
-              router.push("/sites");
-              toast.success(`Successfully deleted site!`);
-            }
-          })
-          .catch((err: Error) => toast.error(err.message))
-      }
+      action={deleteSiteData}
       className="rounded-lg border border-red-600 bg-white dark:bg-black"
     >
       <div className="relative flex flex-col space-y-4 p-5 sm:p-10">
